+++ lib/QECCore/src/QECCore.jl
@@ -15,11 +15,13 @@ export AbstractECC, AbstractQECC, AbstractCECC, AbstractCSSCode, AbstractDistanc
 export Perfect5, Cleve8, Gottesman
 
 # CSS Codes
-
-export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct
+export Toric, Bitflip3, Phaseflip3, Shor9, Steane7, Surface, CSS, QuantumReedMuller, Triangular488, Triangular666, DelfosseReichardt,  QuantumTannerGraphProduct, CyclicQuantumTannerGraphProduct
 
 # Classical Codes
 export RepCode, ReedMuller, RecursiveReedMuller, Golay, Hamming
+
+# utilities
+export search_self_orthogonal_rm_codes
 
 include("interface.jl")
 include("codes/util.jl")
@@ -46,6 +48,9 @@ include("codes/quantum/quantumtannergraphproduct.jl")
 include("codes/classical/reedmuller.jl")
 include("codes/classical/recursivereedmuller.jl")
 include("codes/quantum/quantumreedmuller.jl")
+
+# Delfosse-Reichardt Codes
+include("codes/quantum/delfosse_reichardt_code.jl")
 
 function __init__()
     if isdefined(Base.Experimental, :register_error_hint)